refactor(analysis): extract Python complexity span owner
